score 32bit

+++ hw_webp.h
@@ -1,6 +1,6 @@
 #include <stdint.h>
 
-typedef int64_t score_t;     // type used for scores, rate, distortion
+typedef int32_t score_t;     // type used for scores, rate, distortion
 
 typedef struct VP8Matrix {
   uint16_t q_[16];        // quantizer steps
